requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
docutils==0.16
gunicorn==20.0.4
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
